refactor(js): align with new proto locations
