+++ openfeature-provider/js/src/LocalResolver.ts
@@ -1,5 +1,5 @@
-import type { ResolveWithStickyRequest, ResolveWithStickyResponse } from './proto/resolver/api';
-import type { SetResolverStateRequest } from './proto/messages';
+import type { ResolveWithStickyRequest, ResolveWithStickyResponse } from './proto/confidence/wasm/wasm_api';
+import type { SetResolverStateRequest } from './proto/confidence/wasm/messages';
 
 export interface LocalResolver {
   resolveWithSticky(request: ResolveWithStickyRequest): ResolveWithStickyResponse;
