+++ openfeature-provider/js/src/WasmResolver.test.ts
@@ -1,8 +1,8 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { UnsafeWasmResolver, WasmResolver } from './WasmResolver';
 import { readFileSync } from 'node:fs';
-import { ResolveWithStickyRequest } from './proto/resolver/api';
-import { ResolveReason } from './proto/types';
+import { ResolveWithStickyRequest } from './proto/confidence/wasm/wasm_api';
+import { ResolveReason } from './proto/confidence/flags/resolver/v1/types';
 import { WriteFlagLogsRequest } from './proto/test-only';
 
 const moduleBytes = readFileSync(__dirname + '/../../../wasm/confidence_resolver.wasm');
